# Conflicts:
#	src/main/java/com/uet/k62/web/system/examination/restcontroller/UserController.java
#	src/main/java/com/uet/k62/web/system/examination/service/impl/UserServiceImpl.java
#	src/main/resources/static/JS/Layout.js

+++ src/main/java/com/uet/k62/web/system/examination/model/dtos/QuestionResponseDTO.java
@@ -7,7 +7,7 @@ import java.util.Date;
 
 @Data
 public class QuestionResponseDTO {
-    private BigInteger id;
+    private Integer id;
 //    private int questionTypeId;
     private String questionTypeCode;
     private String content;
